Know we could also handle pid with values higher than 256.

+++ trunk/vdr-1.4.1/PLUGINS/src/mminput/FilterHandle.hpp
@@ -23,7 +23,7 @@ private:
 		cSectionFilter *sf;
 	};
 	
-        int8_t PidNum[MAXDEVICEFILTER];
+        int PidNum[MAXDEVICEFILTER];
 	FilterPids FH[MAXDEVICEFILTER];
 	int maxFilter;
 	bool ClosePipe(int Pid, int tid);
